+++ app/controllers/graders_controller.rb
@@ -87,6 +87,29 @@ class GradersController < ApplicationController
       end
     end
 
+    # It appears that the params[:graders] received is always an array of stringed integers instead of integers.
+    # This is the case both from the React front-end (it sends id's as strings; this should be dealt with in
+    # the future), as well as from the requests generated in the RSPEC tests, where it appears that array elements
+    # (in request parameters) are automatically converted to strings for HTTP transmission.
+    # This map handles the required conversion from stringed integers to integers.
+    grader_ids.map!(&:to_i)
+
+    if %w[assign random_assign].include? params[:global_actions]
+      inactive_graders_hash =
+        current_course.tas.joins(:user).where(roles: { hidden: true }).pluck(:id, :user_name)
+                      .map { |x| [x[0], x[1]] }.to_h
+
+      inactive_graders_for_flash = inactive_graders_hash.select { |k| grader_ids.include? k }
+
+      grader_ids.reject! { |grader_id| inactive_graders_hash.key? grader_id }
+
+      if inactive_graders_for_flash.size > 0
+        flash_now(:error,
+                  I18n.t('groups.invite_member.errors.inactive_grader',
+                         user_names: inactive_graders_for_flash.values.join(', ')))
+      end
+    end
+
     case params[:current_table]
     when 'groups_table'
       grouping_ids = params[:groupings]
@@ -182,6 +205,8 @@ class GradersController < ApplicationController
     Grouping.randomly_assign_tas(grouping_ids, grader_ids, weightings, @assignment)
   end
 
+  # This helper is only expected to be invoked from within the global_actions method, which
+  # filters 'grader_ids' to remove inactive graders (in the cases of assign and random_assign)
   def assign_all_graders(grouping_ids, grader_ids)
     Grouping.assign_all_tas(grouping_ids, grader_ids, @assignment)
   end